blockdata for Search Tree Caves
This is a 20x20 map using tileset 2. There are 3 warps corresponding to
3 cave entrances.

The top left warp is at y=21 x=12. Top right warp is at y=21 x=24. The
bottom warp is at y=27 x=18.

This map would make a really good binary search map, just reuse the
blockdata for the next few maps on the two upper caves and loop it a few
times.